feat: update FE and BE

+++ models/OrderItem.js
@@ -7,9 +7,10 @@ const OrderItem = sequelize.define('OrderItem', {
   orderId: { type: DataTypes.BIGINT.UNSIGNED, allowNull: false, comment: 'ID đơn hàng' },
   printType: { type: DataTypes.ENUM('DOCUMENT', 'PHOTO', 'BANNER'), allowNull: false, defaultValue: 'DOCUMENT' },
   pricingMode: { type: DataTypes.ENUM('PER_PAGE', 'PER_SHEET', 'FIXED'), allowNull: false, defaultValue: 'PER_PAGE' },
-  paperSizeId: { type: DataTypes.BIGINT.UNSIGNED, allowNull: false },
-  colorModeId: { type: DataTypes.BIGINT.UNSIGNED, allowNull: false },
-  sideId: { type: DataTypes.BIGINT.UNSIGNED, allowNull: false },
+  // Khớp với schema: SMALLINT UNSIGNED
+  paperSizeId: { type: DataTypes.SMALLINT.UNSIGNED, allowNull: false },
+  colorModeId: { type: DataTypes.SMALLINT.UNSIGNED, allowNull: false },
+  sideId: { type: DataTypes.SMALLINT.UNSIGNED, allowNull: false },
   pages: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 0 },
   quantity: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 1 },
   setCount: { type: DataTypes.INTEGER, allowNull: true },
